feat(chart): create bare entrypoint to reduce bundle size
